+++ CHANGELOG.md
@@ -740,3 +740,5 @@
 * Define `Probably` pattern synonym
 * Define `yi_yi'_yi'` operator
 * Define `yi_yi_yi'_yi'` operator
+* Define `yi_lm_ds` operator
+* Define `yi_yi_lm_ds` operator